# frontend/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "ConversationCreate": {
        "description": "Start or resolve a 1:1 conversation with another user.",
        "properties": {
          "other_user_id": {
            "description": "Other participant user id",
            "format": "uuid",
            "title": "Other User Id",
            "type": "string"
          }
        },
        "required": [
          "other_user_id"
        ],
        "title": "ConversationCreate",
        "type": "object"
      },
      "ConversationDetail": {
        "description": "Full conversation header for open/view (id + partner).",
        "properties": {
          "booking_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "partner": {
            "$ref": "#/components/schemas/ConversationPartner"
          },
          "partner_last_read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Last Read At"
          },
          "partner_read_up_to_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Read Up To Message Id"
          }
        },
        "required": [
          "conversation_id",
          "partner",
          "created_at"
        ],
        "title": "ConversationDetail",
        "type": "object"
      },
      "ConversationListItem": {
        "description": "One row in the inbox (partner + optional last message preview).",
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "has_unread": {
            "default": false,
            "title": "Has Unread",
            "type": "boolean"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "last_message_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message Preview"
          },
          "partner": {
            "$ref": "#/components/schemas/ConversationPartner"
          }
        },
        "required": [
          "conversation_id",
          "partner"
        ],
        "title": "ConversationListItem",
        "type": "object"
      },
      "ConversationPartner": {
        "description": "Minimal partner profile for conversation lists.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "full_name"
        ],
        "title": "ConversationPartner",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageCreate": {
        "description": "Send a message in a conversation.",
        "properties": {
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "MessageCreate",
        "type": "object"
      },
      "MessageGapResponse": {
        "description": "Reconnect backfill messages since a known message id.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/app__domain__chat__schema__MessageResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "last_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message Id"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "title": "MessageGapResponse",
        "type": "object"
      },
...
      "PaginatedConversationsResponse": {
        "description": "Cursor-paginated inbox (conversation list).",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor — pass as `after` query param.",
            "title": "Next Cursor"
          }
        },
        "title": "PaginatedConversationsResponse",
        "type": "object"
      },
      "PaginatedMessagesResponse": {
        "description": "Cursor-paginated message list.",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/app__domain__chat__schema__MessageResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor — pass as `after` query param.",
            "title": "Next Cursor"
          }
        },
        "title": "PaginatedMessagesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__domain__chat__schema__MessageResponse": {
        "description": "Single message in API responses.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "integer"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "conversation_id",
          "sender_id",
          "body",
          "created_at"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/conversations": {
      "get": {
        "description": "Inbox for the current user with partner info and last message (cursor pagination).",
        "operationId": "list_conversations_api_v1_chat_conversations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Cursor from previous page next_cursor",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor from previous page next_cursor",
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedConversationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "רשימת השיחות שלי",
        "tags": [
          "Chat",
          "Chat"
        ]
      },
      "post": {
        "description": "Get or create a 1:1 thread with other_user_id.\nReturns conversation_id and partner details.",
        "operationId": "create_or_get_conversation_api_v1_chat_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "פתיחת שיחה (או קבלת שיחה קיימת)",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/by-booking/{booking_id}": {
      "post": {
        "description": "Get or create a driver–passenger thread for a booking.\nOnly booking parties may open; status must be pending_approval or confirmed.",
        "operationId": "create_or_get_conversation_by_booking_api_v1_chat_conversations_by_booking__booking_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "פתיחת שיחה דרך booking",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}": {
      "get": {
        "description": "Single conversation header; participant only.",
        "operationId": "get_conversation_api_v1_chat_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "פרטי שיחה",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/calendar.ics": {
      "get": {
        "description": "Export conversation to iCal (.ics); requires AI analysis (not fully implemented).",
        "operationId": "export_conversation_calendar_api_v1_chat_conversations__conversation_id__calendar_ics_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "ייצוא שיחה ללוח שנה (iCal)",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/messages": {
      "get": {
        "description": "Message history with cursor pagination; participants only.",
        "operationId": "list_conversation_messages_api_v1_chat_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from previous page next_cursor",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from previous page next_cursor",
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "היסטוריית הודעות",
        "tags": [
          "Chat",
          "Chat"
        ]
      },
      "post": {
        "description": "Post a message; participants only. Rate limited: 30 messages/minute per user.\n\nOptional Idempotency-Key: same semantics as POST request-ride-from-search (Redis TTL;\nmismatch → 422, in-flight → 409 + Retry-After).",
        "operationId": "post_message_api_v1_chat_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "description": "Optional UUID per send intent — dedupe double-submit / client retries (same key + same body replays cached 201).",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional UUID per send intent — dedupe double-submit / client retries (same key + same body replays cached 201).",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__domain__chat__schema__MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "שליחת הודעה",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/messages/gap": {
      "get": {
        "description": "Reconnect backfill for messages newer than since_message_id.",
        "operationId": "list_conversation_messages_gap_api_v1_chat_conversations__conversation_id__messages_gap_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_message_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Since Message Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageGapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "פער הודעות מאז message_id אחרון",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/read": {
      "post": {
        "operationId": "mark_read_api_v1_chat_conversations__conversation_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
    "/api/v1/chat/unread-count": {
      "get": {
        "operationId": "unread_count_api_v1_chat_unread_count_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Unread Count Api V1 Chat Unread Count Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unread Count",
        "tags": [
          "Chat",
          "Chat"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://127.0.0.1:8000"
    }
  ]
}